fix: correct float type and consolidate test user helper
- CompoundConstraintValueValidator: change $time property from int to
  float since microtime() returns float
- Consolidate duplicated temp account user logic into
  TestEnvironment::getTestUser() — previously repeated in PageCreator,
  PageEditor, PageDeleter, and XmlImportRunner

+++ src/DataValues/ValueValidators/CompoundConstraintValueValidator.php
@@ -23,7 +23,7 @@ class CompoundConstraintValueValidator implements ConstraintValueValidator {
 	 */
 	private $hasConstraintViolation = false;
 
-	private int $time = 0;
+	private float $time = 0;
 	private int $count = 0;
 	private $contextPage = '';
